First push
This bot is written in Python. It was copied from an open source and I have modified it in 2018. The main files in cogs folder were not created by me but some of them were modified by me to fit my discord server. I did create some cogs too, it was in early 2018 I don't exactly remember what I have used in the open source and what I have created.